(test) Integration test: installing dependencies
Currently only the individual modules (e.g. Update, Install) are covered
by tests. In order to refactor, we need some test coverage at a higher
level. This is a fairly complete integration test: it creates actual
files in a directory.

It's based on `tests/plugins_test.php` which tested the (now deprecated)
`whippet plugins install` command.

NOTE: This sort of test interacts with the filesystem and makes real
calls out to git repos and APIs - they will be slow.