(feature) Real implementation of BaseApi
Since we're only doing GET requests, maybe this is sufficient for now?

Note there are no tests: the purpose of this object is to provide an
interface to methods which talk over http so that we can mock them out
in tests, and seamlessly change the implementation without needing to
touch any other code.

Note that we _can't_ easily unit test this, because of the call to
`file_get_contents`, which is a core PHP function.

Hopefully it's trivial enough that we can easily imagine all possible
paths through it, so tests wouldn't add any value, even if we were able
to mock out that method.

If this class becomes any less trivial in future, then we should test
it, and extract another trivial base-level object.