(feature) Warning message is prominent
When there are no inspections for a given plugin, it's important that
the user takes some sort of action.

It's annoying that this Heredoc is duplicated in all of these tests, but
if we want to have our unit test mocks responding with realistic data
then I think it's worthwhile.