Refactor updater
Removes code duplication, consolidates core updating tasks under new
update() method, which is passed an array of dependencies to update.

Whippet.json is now only loaded in updateSingle() after testing for
the presence of whippet.lock, and that the dependency has been provided
in the correct format. Therefore,
`testUpdateSingleIncorrectFormat()` and `testUpdateSingleWithNoLock()`
no longer need to generate a whippet.json file, and that section of those
tests has been removed.