(refactor) Inject inspections api path
Now this is just another part of the configuration. Arguably there's now
no longer much value in passing in the host and path separately, except
that these will change for different reasons, and e.g. in development we
want to be able to change the host but keep the path the same.