(feature) Service for fetching file contents
A very thin wrapper around the php functions which fetch file contents.
This decouples the rest of the code from knowing exactly know we
fetch file contents (and handle file-fetching errors), and will allow us
to stub out interactions with the filesystem in tests.

This is untested, but should be trivial enough that it's clear what it
does.

If it were to get any more complicated then we should extract the
complicated logic and test that, injecting a simple file-fetching
object.