WIP - service for loading WhippetJson
The current way (calling static methods, building classes from strings,
using factories) is extremely complicated in relation to how simple the
task is.

It's annoying here to have to wrap the file in a Request object to pass
it up to the next level. maybe better to just fail here and return the
file otherwise.