(refactor) extract inspectionDetailsMessage
There are a few conditional branches here for something which isn't the
main point of the method, so it makes sense to extract a method.

Complication: I'd like to keep all the `echo()` calls at the same
level, so we need to handle the case where we don't want to output
anything. Options for that are to return `null`, `false` or `''` - none
of which are ideal. `null` feels the most expressive of the options.

This extracted method is too long and has a nested conditional, but I
can't really see a way of improving it :/