(refactor) inject the lockFile into installer
Lockfiles are used in multiple actions, each of which are responsible
for fetching them individually. This moves lockfile fetching up to the
next level.

One test has been deleted: it relates to a test which is now no longer
done inside this class, but is done higher up and tested elsewhere.

Note that in `update` the lockfile needs to get fetched _after_ update
has happened, since update modifies the lockfile.