(refactor) GitIgnore returns [] if file not exists
This feels more like a responsibility of the GitIgnore class than of its
callers.

As a result, from the perspective of the Updater class there is no
difference between an empty gitignore and a missing one. I feel like
that's correct: there's no reason that it should be concerned with the
state of the filesystem.

The .gitignore file was referenced both in the GitIgnore class (to get
its contents) and in the Updater class (to check whether it exists).

Isolating in one place the knowledge of what this file is called and how
to interact with it will allow us to significantly simplify some of the
code which calls Gitignore.

Previously the testing approach was to touch an empty .gitignore file,
purely in order to make a conditional pass, the result of which was that
either an empty array was returned, or a call to
`GitIgnore::get_ignores`.

Moving this check inside GitIgnore removes another place where the
Updater class is aware of the filesystem.

A couple of test assertions change as a result: before we could fail to
`get_ignores` because of either a nonexistent .gitignore file OR because
of an error before we got to that point. Now only the latter is
possible.