(refactor) extract a dependency object
Responsible for handling default values

Previously the dependency existed as an associative array, and was
passed around along with associated information like type. This refactor
embodies the Dependency concept as an actual object.

This helps us push $sources right up to the top level and cleans up the
code for setting default values for `ref` and `src`, which are concerns
of that object.

The handling of src still isn't great. The problem is that we need to
handle two legitimate cases...:

  1 a source is explicitly specified for a dependency
  2 a source is not explicitly specified, but a default repo is declared
    in the whippet.json, to be used for all dependencies of that type

...and one error case:

  3 Neither a specific source, nor a default source are specified.

I feel like a better solution might ultimately be to do a validation
step after loading the json file (but before doing anything else) to
check that a source can be built for everything, and additionally that
every dependency has a name.