(fix) Don't write if no dependencies
Previously we were writing a lockfile and gitignore even if there were
no dependencies.

Looks like this $count variable approach was only needed because of the
nested loop that we just got rid of.

Now that we have a flat array it's a lot simpler to just check if it has
any elements, and exit early if not.

Also note that we only need to check for this in the updateAll case,
since for updateSingle we're creating an array of one element ourselves.