(refactor) pass lockfile and locked deps as args
Rather than fetching the locked dependencies deep down the call stack,
isolate this interaction with the lockfile up at the top level function.

This is a step towards injecting the lockfile in the constructor.

Note that in the case of a single update we only need to pass the single
locked dependency to the update function.

We don't need to check that this single dependency already exists in the
lockfile: if we are newly adding this dependency then it _won't_ exist
in the lockfile yet.