(refactor) clarify gitignore interactions
This is an attempt to expose the current approach, which is roughly:

1. load ignores from gitignore
2. look through the lockfile and delete anything in there from the
gitignore
3. add back in dependencies from the JSON file
4. save the updated list of ignores back to the file

This was previously obfuscated by loadGitignore doing a bit too much,
and the fact that ignores was passed around as a property, so it was
hard to see where and in what order it was being written to and read.

This refactoring includes removing $ignores as an instance variable.
Instance variables initialized outside of the constructor result in more
complicated flow through the code, since they can potentially be
modified in any function.

The refactored code takes a functional approach: passing the ignores
array to a function which returns an updated array as its return value.
In this way we avoid mutating state.