(fix test) only mock an empty gitignore if tested
These specs were implicitly testing the scenario where the gitignore
file is initially empty. We have a specific test for that case, so we
should eliminate this variance from the other tests.

getGititgnore both _stubs_ calls returning the set of ignores (the first
argument) and _mocks_ calls writing the set of ignores (the second
argument).

In these two tests we are updating a single plugin, so we don't expect
any change in the gitignore.