(fix) Update single dep only ignores that dep
The existing code said "even if we're only updating a single dependency,
loop through all the locked dependencies and create a gitignore based on
those.

If we have update_single only knowing about the single dependency which
we intend to update, then the code can become a lot cleaner.

Arguably if the gitignore is not up-to-date for other dependencies, then
we're in an exceptional scenario and updateSingle isn't the right tool
to recover from this.