(feature) Store the hash of the json file
Currently we're using an object to get at the data in whippet.json, but
always getting the hash by accessing the file directly, which leads to a
bit of duplication, and in the case of Updater, reading the file twice:
once to load it and get at the data, then separately to get the hash.

This commit calculates the hash at the same point at which JSON data is
parsed.

This does mean that accessing the hash requires parsing the data in the
JSON file (and blowing up if it's invalid) but that doesn't seem
unreasonable.