(feature) Disable inspections check from cli
The plan is to limit access to the inspections API - e.g. by ip-locking
it to dxw addresses. This means that for non-dxw employees, and people
working away from their normal locations, the API won't be available,
which will slow down the installation step and show errors in the
output.

This PR adds a command-line flag which disables the calls to the API.

In order to use polymorphism (passing an InspectionChecker or
NullInspectionChecker which have the same interface) I've removed the
type declaration. An alternative would be to create an abstract class
which both classes implement, and use that in the type declaration, but
that feels like overengineering.