(feature) 'Generate' clones the whippet-theme-template repo
The problem with having 'generate' copy files from a submodule is that
the submodule can easily get out date and the developer may not know there
has been an update for it.

This PR will clone down the latest `whippet-theme-template` in to the
correct directory.

I had to change the unlink to a system call to remove the .git folder
as there was an issue with PHP not having the permission to remove the
cloned dir.

Also included a simple function to rename the namespace of the newly
created theme to a camel cased version of the directory name

In a future PR we will include a setup script that does `composer
install` and all the other commands to get up and running