(chore) Steamlined generate theme code
As pointed out in code review:

It's better to use the https version of the whippet-theme repo as people
outside the company may be using whippet.

We have a recursive delete method we can use instead of a system call.
It works well.

Removed .gitmodules file as there is no longer any submodules in the
project.