Use Wordpress-template repo as source for theme generation
Before: `whippet generate theme` used the Whippet Theme Template repo as the source for the generated theme. That repo is being replaced by the WordPress Template, which includes a theme within it

Now: `whippet generate theme` uses the theme from the WordPress Template as its source.

Note, because the new repo uses `Theme/` as the namespace for the theme, this means the auto-replacement of namespacing that whippet theme generation used to do is no longer possible.

Resolves: https://github.com/dxw/whippet/issues/151