Use the template's root .gitignore on theme generation
We have consolidated the wordpress-template repo's .gitignore rules into
a single .gitignore file, in the root of the template.

To allow for this, when whippet generates a theme, it will now copy
that root .gitignore into the theme folder by default.

If you set the -nogitignore option when generating a theme (e.g. if you
are generating a theme within a whippet app that already has a root
.gitignore), no gitignore file will be generated.