Make guzzle 6.5.8 the minimum accepted version
Although we'd already updated our `composer.lock` file to require a
minimum of guzzlehttp 6.5.8, when a package is included as a composer
dependency in another project, the requirements in the `composer.lock`
file are ignored, and only those in the `composer.json` are considered.

Therefore, to make sure that the guzzlehttp vulnerability in versions
below 6.5.8 is patched in projects where Whippet is included as a
composer dependency, let's bump the version in `composer.json` as well.