Add version commmand
All packages installed with brew have the version of the package in the
name of the parent folder. Whereas those installed from source have the
git tags present.

With that in mind, either the git tags or parent folder name is referenced to
print the release version used.

Relates: https://github.com/dxw/whippet/issues/159