Add `whippet deps describe` command
This will output a JSON format listing of your dependencies, and what
versions they correspond to (based on the tags in the git repos for
those dependencies).

This is useful because Whippet doesn't have any inherent understanding
of versioning - it either pins to a specific git reference, or updates to
that latest commit available on `main` or `master`. This command allows
a developer to quickly see what versions Whippet will install. It
doesn't need to install the dependencies, or use the GitHub API to do
so, it just uses `git ls-remote` and filters the result. So it should be
a) reasonably quick; and b) possible to run in CI without worrying about
GitHub API rate limiting.