Allow for annotated tags
Sometimes git ls-remote can return a tag in the format of
`[tag-name]^{}`. This indicates that while that tag is attached to the
commit, the tag itself exists as a separate git object, and is usually
indicative of an annotated tag.

We don't need to worry about if the tag exists as a separate object - we
just care about the commit hash it's attached to, so when we return the
tag reference we want to remove the `^{}` notation and just return the
tag name itself.