Generated app can be deployed
Whippet expects that a deployable app is a Git repo, with
at least one commit. Previously, an app generated by Whippet
was a Git repo, but did not have any commits, and so could
not be deployed.

This commit adds a single commit to generated apps, to fix that.

Fixes #47