Dependency types are not hard-coded literals
Previously we hard-coded the list of dependency types wherever
we needed to use them. In order to manage language packs within
Whippet, we now need to add to the list of dependency types.
So, rather than have them scattered over the code, we now have
them in one place.

Once we have dropped PHP 7.4 support, we should really use an
enumeration here.

Fixes #70