Add languages dependency type
Note that at this point, it isn't clear how we will need to
manage language packs, and it may be that we do not pin
languages to versions or hashes in the lockfile. For this
reason, the relevant test data is mostly blank, with the
expectation that this will change (or more tests will be
added) in later commits.