Refactor describer
In order to manage language files, we will need a semver
version for each plugin and theme in the whippet.json file.
These versions are obtainable via the Describer class, but
currently the relevant method just prints out version
information.

This commit splits that code into two parts, one to fetch
and return the information, and one to print it, so that
we can reuse the code elsewhere. Since both methods are already
tested by the existing tests, this commit doesn't add any more.