Add language packs to whippet.lock files
Language packs in WordPress are available for WP Core, plugins
and themes. In a manual install, they can be added with:

    wp language core install
    wp language plugin classic-editor ja
    wp language theme install twentytwenty ja

and the wp cli command will work out the required versions of
the language packs, based on already installed versions of the
core, plugins and themes.

The language packs themselves are downloaded a .zip files from
wordpress.org.

In this commit, we replicate this behaviour, by:

* Looking up the versions of language packs that are required
  for each langauge or theme, based on the APIs documented here:
  https://codex.wordpress.org/WordPress.org_API#Translations
* We add the locations of the relevant .zip files to the 'src'
  key/value pair in the whippet.lock file with the revision
  of the dependency (core/plugin/theme) that the language
  pack relates to.

Note that language packs are different to plugins and themes
in a number of ways. Relevant to the Updater, is the fact that
language packs relate to other dependencies. So, the 'name'
value for each one is either:

* Just the language code (e.g. 'en_US') if the related
  dependency is core.
* '<language>/<type>/<slug>' otherwise, e.g.
  'ja/plugins/classic-editor' or 'ja/themes/twentytwenty'.