Validate language packs in lockfiles
Language packs are slightly different to other dependencies
because a single language pack may have separate sources
for WP core, any plugins for which the language is available
and any themes.

We expect that if a language code is in the .json file, it
should have a related language pack for WP core as a minimum,
but there may be any number more entries in the lockfile for
that language code.