Describe language packs in lockfiles
This commit ensures that `whippet deps describe`
shows information about language packs. Note that
the version numbers shown are the versions of the
related dependencies, so for example output like:

    "languages": {
        "ja": "6.3.1",
        "ja/plugins/classic-editor": "1.6.3",
    }

is showing the versions of WordPress Core, and the
Classic Editor plugin. Versions of the translations
themselves are not when installing language packs,
but translations for WP Core, plugins and themes need
to be consistent with the the versions of those
dependencies that Whippet is installing. Therefore,
we don't need to store version informations for
language packs themselves.