Add return type will change attribute to toString method
From PHP 8.1 up, internal methods like `__toString()` that can be
over-ridden require the over-riding method to declare a compatible
return type. Failure to do so will result in a deprecation notice.

However, currently, if we declare a return type on this method, it
results in the Kahlan tests failing, as the `__toString()` method Kahlan
generates on its double for this class does not have a return type, so
is viewed as incompatible with the class it's trying to double.

Declaring the ReturnTypeWillChange attribute prevents the deprecation
notice being issued, and stops Psalm complaining about it.

See: https://php.watch/versions/8.1/ReturnTypeWillChange