Allow for the fact $this->plugins_locked is an object
Before: we were trying to loop through `$this->plugins_locked` as if it
was an array (as previously we'd been using `foreach` with key/value
pairs). However, it's actually an object, and foreach-ing an object
produces a key-value array of the visible properties of that object.

Now: we use get_object_vars to get the visible properties as an array,
then use array_keys on that.