WordPress Core versions check semver constraints
In repos we have a config file that determines the version of
WordPress Core that needs to be deployed to a live site.
This can be a full semver version (e.g. 1.2.3) or a "version"
that is actually a moveable tag related to a major release
(e.g. v6).

Previously, we did not check for this edge case, which means
that in some cases, we did not return a full semver version
and this caused a bug in updating language packs whereby
old language packs would be written into repos and lockfiles.

Now, we remove any leading 'v' characters and return a fully
qualified semver version (e.g. 1.2.3).

We make an assumption that the list of available core versions
(e.g. https://api.wordpress.org/translations/core/1.0/?version=6)
is ordered by most recent release first.