Run 'git fetch' with '--force' if no revision given
When we run 'git fetch' without a specific revision we
usually run it with '--all' which fetches tags (among
other things).

We now have movable tags in our plugin repositories
and when a tag is moved a 'git fetch' or 'checkout'
will produce a 'would clobber existing tag' error on
the command line, which warns the user that the tag
that has already been checked into the local clone
will be overwritten.

This commit fixes that error by running 'fetch' with
'--force' if we do not specify a revision.