Warn user if they pull from an archived GitHub repo
This commit issues a warning to the user if they clone or checkout from an
archived GitHub repository. In dxw, archiving a GitHub repository containing
a dependency is likely to mean that the repository will soon be removed, and
so we need to make users aware of this. However, we do not want to cause
an error here, because we do not know whether Whippet is being run in the
context of a local development environment or as part of a deploy process.

Note that because we are only enabling this feature for a specific dxw use case,
this commit does not make the effort to abstract out GitHub-specific code in
a way that would make it easy to generalise the feature for GitLab, BitBucket,
etc. This can be done in a future PR, if anyone needs it, but ideally we would
want to improve the unit tests before hand.