Allow apps to deploy different WP Core versions on prod
Previously we deployed the same WP Core version on all
environments. To descibe the version of Core that we
wish to deploy we use a movable tag, which is usually
the major version tag of the latest release, e.g. 'v6'.

However, this means that for patch and minor version
updates, moving the movable tag will change deploys
on both staging and production.

This commit changes that strategy, and if there is an
no environment variable defined that looks like it
describes a production environment, we try to deploy
to a revision called 'latest' and then fall back to
the revision from the JSON file if that does not exist.

Environment variables that might describe a production
environment are:

* 'WP_ENVIRONMENT_TYPE' - found in wp-config.php
* 'WP_CONFIG_WP_ENVIRONMENT_TYPE' - specific to dxw